Logica concluida / sem estilização
