--- Madson/Components/Confirmacao.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Confirmacao = () => {
-    return (
-        <div>
-        
-        </div>
-    )
-}
-
-export default Confirmacao;
