--- Madson/Components/PedidoEmAndamento.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const PedidoEmAndamento = () => {
-    return (
-        <div>
-        
-        </div>
-    )
-}
-
-export default PedidoEmAndamento;
