--- Madson/Pages/Carrinho.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Carrinho = () => {
-    return (
-        <div>
-        
-        </div>
-    )
-}
-
-export default Carrinho;
