--- Madson/Pages/Perfil.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Perfil = () => {
-    return (
-        <div>
-        
-        </div>
-    )
-}
-
-export default Perfil;
