+++ four-food/src/Pages/Restaurante/components/RestauranteInfo.jsx
@@ -1,7 +1,82 @@
 import React from 'react'
-import { Card, CardContent, CardMedia, Typography } from '@mui/material'
 import styled from 'styled-components'
 
+const StyledContainer = styled.div`
+	display: flex;
+	flex-direction: column;
+	width: 22.5rem;
+`
+const StyledImg = styled.img`
+	width: 20.5rem;
+	height: 7.5rem;
+	margin: 1.063rem 1rem 0.75rem;
+	object-fit: contain;
+`
+const StyledNome = styled.span`
+	width: 20.5rem;
+	height: 1.125rem;
+	margin: 0.75rem 1rem 0.5rem;
+	font-family: Roboto;
+	font-size: 1rem;
+	font-weight: normal;
+	font-stretch: normal;
+	font-style: normal;
+	line-height: normal;
+	letter-spacing: -0.39px;
+	color: #e8222e;
+`
+const StyledCategoria = styled.p`
+	width: 6.5rem;
+	height: 1.125rem;
+	margin: 0.5rem 0.5rem 0.5rem 1rem;
+	font-family: Roboto;
+	font-size: 1rem;
+	font-weight: normal;
+	font-stretch: normal;
+	font-style: normal;
+	line-height: normal;
+	letter-spacing: -0.39px;
+	color: #b8b8b8;
+`
+const StyledEntrega = styled.span`
+	width: 6.5rem;
+	height: 1.125rem;
+	margin: 0.5rem 0.5rem 0.625rem 1rem;
+	font-family: Roboto;
+	font-size: 1rem;
+	font-weight: normal;
+	font-stretch: normal;
+	font-style: normal;
+	line-height: normal;
+	letter-spacing: -0.39px;
+	color: #b8b8b8;
+`
+const StyledFrete = styled.span`
+	width: 6.5rem;
+	height: 1.125rem;
+	margin: 0.5rem 8rem 0.625rem 0.5rem;
+	font-family: Roboto;
+	font-size: 1rem;
+	font-weight: normal;
+	font-stretch: normal;
+	font-style: normal;
+	line-height: normal;
+	letter-spacing: -0.39px;
+	color: #b8b8b8;
+`
+const StyledEnderoco = styled.p`
+	width: 20.5rem;
+	height: 1.125rem;
+	margin: 0.625rem 1rem 1rem;
+	font-family: Roboto;
+	font-size: 1rem;
+	font-weight: normal;
+	font-stretch: normal;
+	font-style: normal;
+	line-height: normal;
+	letter-spacing: -0.39px;
+	color: #b8b8b8;
+`
 const RestauranteInfo = ({
 	nome,
 	endereco,
@@ -11,38 +86,31 @@ const RestauranteInfo = ({
 	imagem,
 }) => {
 	const InfoEntrega = styled.div`
-		.tempo-de-entrega {
-			margin-right: 25px;
-			height: 1.125rem;
-			font-family: Roboto;
-			font-size: 1rem;
-			font-weight: normal;
-			font-stretch: normal;
-			font-style: normal;
-			line-height: normal;
-			letter-spacing: -0.39px;
-			color: var(--greyish);
-		}
+		margin-right: 25px;
+		height: 1.125rem;
+		font-family: Roboto;
+		font-size: 1rem;
+		font-weight: normal;
+		font-stretch: normal;
+		font-style: normal;
+		line-height: normal;
+		letter-spacing: -0.39px;
+		color: #b8b8b8;
 	`
 
 	return (
 		<>
-			<Card sx={{ maxWidth: 345 }}>
-				<CardMedia component='img' image={imagem} alt='restaurante Logo' />
-				<CardContent>
-					<Typography gutterBottom variant='h5' component='div'>
-						{nome}
-					</Typography>
-					<Typography variant='body2' color='text.secondary'>
-						<p>{categoria}</p>
-						<InfoEntrega>
-							<span className='tempo-de-entrega'>{tempoDeEntrega} min</span>
-							<span className='tempo-de-entrega'>Frete: R${frete}</span>
-						</InfoEntrega>
-						<p>{endereco}</p>
-					</Typography>
-				</CardContent>
-			</Card>
+			<StyledContainer>
+				<StyledImg src={imagem} alt='Logo do Restaurante' />
+				<StyledNome>{nome}</StyledNome>
+
+				<StyledCategoria>{categoria}</StyledCategoria>
+				<InfoEntrega>
+					<StyledEntrega>{tempoDeEntrega} min</StyledEntrega>
+					<StyledFrete>Frete: R${frete}</StyledFrete>
+				</InfoEntrega>
+				<StyledEnderoco>{endereco}</StyledEnderoco>
+			</StyledContainer>
 		</>
 	)
 }
